requirements.txt
blis==0.4.1
boto==2.49.0
boto3==1.14.17
botocore==1.17.17
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cymem==2.0.3
docutils==0.15.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
gensim==3.8.3
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-speech==1.3.2
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
MarkupSafe==1.1.1
murmurhash==1.0.2
numpy==1.19.0
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
scipy==1.5.1
six==1.15.0
smart-open==2.1.0
sox==1.4.0
spacy==2.3.1
SQLAlchemy==1.3.18
srsly==1.0.2
thinc==7.4.1
tqdm==4.47.0
urllib3==1.25.9
wasabi==0.7.0
Werkzeug==1.0.1
zipp==3.1.0
